Created utilities file and added test utilities
